+++ src/services/sanity.js
@@ -25,8 +25,6 @@ import { processMetadata, typeWithSortOrder } from '../contentMetaData.js'
 import { globalConfig } from './config.js'
 
 import {
-  fetchCompletedChallenges,
-  fetchOwnedChallenges,
   fetchNextContentDataForParent,
   fetchHandler,
 } from './railcontent.js'
@@ -40,7 +38,7 @@ import {fetchRecentActivitiesActiveTabs} from "./userActivity.js";
  *
  * @type {string[]}
  */
-const excludeFromGeneratedIndex = ['handleCustomFetchAll', 'fetchRelatedByLicense']
+const excludeFromGeneratedIndex = ['fetchRelatedByLicense']
 
 /**
  * Fetch a song by its document ID from Sanity.
@@ -570,12 +568,13 @@ export async function fetchContentRows(brand, pageName, contentRowSlug)
   if (pageName === 'songs') pageName = 'song'
   const rowString = contentRowSlug ? ` && slug.current == "${contentRowSlug.toLowerCase()}"` : ''
   const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`).buildFilter()
+  const childFilter = await new FilterBuilder('', {isChildrenFilter: true}).buildFilter()
   const query = `*[_type == 'recommended-content-row' && brand == '${brand}' && type == '${pageName}'${rowString}]{
     brand,
     name,
     'slug': slug.current,
-    'content': content[]->{
-        'children': child[]->{ 'id': railcontent_id, 'children': child[]->{'id': railcontent_id}, },
+    'content': content[${childFilter}]->{
+        'children': child[${childFilter}]->{ 'id': railcontent_id, 'children': child[${childFilter}]->{'id': railcontent_id}, },
         ${getFieldsForContentType('tab-data')}
         'lesson_count': coalesce(count(*[${lessonCountFilter}]), 0),
     },
@@ -633,21 +632,6 @@ export async function fetchAll(
     progress = 'all',
   } = {}
 ) {
-  let customResults = await handleCustomFetchAll(brand, type, {
-    page,
-    limit,
-    searchTerm,
-    sort,
-    includedFields,
-    groupBy,
-    progressIds,
-    useDefaultFields,
-    customFields,
-    progress,
-  })
-  if (customResults) {
-    return customResults
-  }
   let config = contentTypeConfig[type] ?? {}
   let additionalFields = config?.fields ?? []
   let isGroupByOneToOne = (groupBy ? config?.relationships?.[groupBy]?.isOneToOne : false) ?? false
@@ -670,9 +654,7 @@ export async function fetchAll(
   } else {
     typeFilter = type
       ? `&& _type == '${type}'`
-      : progress === 'in progress' || progress === 'completed'
-        ? " && (_type != 'challenge-part' && _type != 'challenge')"
-        : ''
+      : ''
   }
 
   // Construct the search filter
@@ -768,152 +750,6 @@ export async function fetchAll(
   })
 
   return fetchSanity(query, true)
-}
-
-/**
- * Fetch all content that requires custom handling or a distinct external call
- * @param {string} brand - The brand for which to fetch content.
- * @param {string} type - The content type to fetch (e.g., 'song', 'artist').
- * @param {Object} params - Parameters for pagination, filtering, sorting, and grouping.
- * @param {number} [params.page=1] - The page number for pagination.
- * @param {number} [params.limit=10] - The number of items per page.
- * @param {string} [params.searchTerm=""] - The search term to filter content by title or artist.
- * @param {string} [params.sort="-published_on"] - The field to sort the content by.
- * @param {Array<string>} [params.includedFields=[]] - The fields to include in the query.
- * @param {string} [params.groupBy=""] - The field to group the results by (e.g., 'artist', 'genre').
- * @param {Array<string>} [params.progressIds=undefined] - An array of railcontent IDs to filter the results by. Used for filtering by progress.
- * @param {boolean} [params.useDefaultFields=true] - use the default sanity fields for content Type
- * @param {Array<string>} [params.customFields=[]] - An array of sanity fields to include in the request
- * @param {string} [params.progress="all"] - An string representing which progress filter to use ("all", "in progress", "complete", "not started").
- * @returns {Promise<Object|null>} - The fetched content data or null if not found.
- */
-async function handleCustomFetchAll(
-  brand,
-  type,
-  {
-    page = 1,
-    limit = 10,
-    searchTerm = '',
-    sort = '-published_on',
-    includedFields = [],
-    groupBy = '',
-    progressIds = undefined,
-    useDefaultFields = true,
-    customFields = [],
-    progress = 'all',
-  } = {}
-) {
-  if (type === 'challenge') {
-    if (groupBy === 'completed') {
-      const completedIds = await fetchCompletedChallenges(brand, page, limit)
-      return fetchAll(brand, type, {
-        page,
-        limit,
-        searchTerm,
-        sort,
-        includedFields,
-        groupBy: '',
-        progressIds: completedIds,
-        useDefaultFields,
-        customFields,
-        progress,
-      })
-    } else if (groupBy === 'owned') {
-      const ownedIds = await fetchOwnedChallenges(brand, page, limit)
-      return fetchAll(brand, type, {
-        page,
-        limit,
-        searchTerm,
-        sort,
-        includedFields,
-        groupBy: '',
-        progressIds: ownedIds,
-        useDefaultFields,
-        customFields,
-        progress,
-      })
-    } else if (groupBy === 'difficulty_string') {
-      return fetchChallengesByDifficulty(
-        brand,
-        type,
-        page,
-        limit,
-        searchTerm,
-        sort,
-        includedFields,
-        groupBy,
-        progressIds,
-        useDefaultFields,
-        customFields,
-        progress
-      )
-    }
-  }
-  return null
-}
-
-async function fetchChallengesByDifficulty(
-  brand,
-  type,
-  page,
-  limit,
-  searchTerm,
-  sort,
-  includedFields,
-  groupBy,
-  progressIds,
-  useDefaultFields,
-  customFields,
-  progress
-) {
-  let config = contentTypeConfig['challenge'] ?? {}
-  let additionalFields = config?.fields ?? []
-
-  // Construct the search filter
-  const searchFilter = searchTerm
-    ? groupBy !== ''
-      ? `&& (^.name match "${searchTerm}*" || title match "${searchTerm}*")`
-      : `&& (artist->name match "${searchTerm}*" || instructor[]->name match "${searchTerm}*" || title match "${searchTerm}*" || name match "${searchTerm}*")`
-    : ''
-
-  // Construct the included fields filter, replacing 'difficulty' with 'difficulty_string'
-  const includedFieldsFilter = includedFields.length > 0 ? filtersToGroq(includedFields) : ''
-
-  // limits the results to supplied progressIds for started & completed filters
-  const progressFilter = await getProgressFilter(progress, progressIds)
-
-  let fields = useDefaultFields
-    ? customFields.concat(DEFAULT_FIELDS, additionalFields)
-    : customFields
-  let fieldsString = fields.join(',')
-
-  const lessonsFilter = `_type == 'challenge' && brand == '${brand}' && ^.name == difficulty_string ${searchFilter} ${includedFieldsFilter} ${progressFilter}`
-  const lessonsFilterWithRestrictions = await new FilterBuilder(lessonsFilter).buildFilter()
-
-  const query = `{
-      "entity": [
-        {"name": "All"},
-        {"name": "Novice"},
-        {"name": "Beginner"},
-        {"name": "Intermediate"},
-        {"name": "Advanced"},
-        {"name": "Expert"}]
-          {
-            'id': 0,
-            name,
-            'all_lessons_count': count(*[${lessonsFilterWithRestrictions}]._id),
-            'lessons': *[${lessonsFilterWithRestrictions}]{
-                ${fieldsString},
-                name
-            }[0...20]
-          },
-          "total": 0
-        }`
-  let data = await fetchSanity(query, true)
-  data.entity = data.entity.filter(function (difficulty) {
-    return difficulty.lessons.length > 0
-  })
-  return data
 }
 
 async function getProgressFilter(progress, progressIds) {
@@ -1414,7 +1250,8 @@ export async function fetchLessonsFeaturingThisContent(railcontentId, brand, cou
  */
 async function fetchRelatedByLicense(railcontentId, brand, onlyUseSongTypes, count) {
   const typeCheck = `@->_type in [${arrayJoinWithQuotes(SONG_TYPES)}]`
-  const typeCheckString = onlyUseSongTypes ? `${typeCheck}` : `!(${typeCheck})`
+  let typeCheckString = `@->brand == '${brand}' && `
+  typeCheckString += onlyUseSongTypes ? `${typeCheck}` : `!(${typeCheck})`
   const contentFromLicenseFilter = `_type == 'license' && references(^._id)].content[${typeCheckString} && @->railcontent_id != ${railcontentId}`
   let filterSongTypesWithSameLicense = await new FilterBuilder(contentFromLicenseFilter, {
     isChildrenFilter: true,
@@ -1431,9 +1268,8 @@ async function fetchRelatedByLicense(railcontentId, brand, onlyUseSongTypes, cou
       "related_by_license" :
           *[${filterSongTypesWithSameLicense}]->{${queryFields}}|order(published_on desc, title asc)[0...${count}],
       }[0...1]`
-
   const results = await fetchSanity(query, false)
-  return results['related_by_license'] ?? []
+  return results ? results['related_by_license'] ?? [] : []
 }
 
 /**
@@ -1635,7 +1471,7 @@ export async function fetchLiveEvent(brand, forcedContentId = null) {
  *
  * @example
  * fetchPackData(404048)
- *   .then(challenge => console.log(challenge))
+ *   .then(pack => console.log(pack))
  *   .catch(error => console.error(error));
  */
 export async function fetchPackData(id) {
@@ -2030,6 +1866,9 @@ export async function fetchSanity(
         console.log('fetchSanity Results:', result)
       }
       let results = isList ? result.result : result.result[0]
+      if (!results) {
+        throw new Error('No results found')
+      }
       results = processNeedAccess
         ? await needsAccessDecorator(results, userPermissions, isAdmin)
         : results