+++ src/services/user/profile.js
@@ -39,11 +39,5 @@ export async function otherStats(userId = globalConfig.sessionConfig.userId) {
  */
 export async function deleteProfilePicture() {
   const url = `${baseUrl}/v1/users/profile_picture`
-  const response = await fetchHandler(url, 'DELETE')
-
-  if (!response.ok) {
-    const problemDetails = await response.json()
-    console.log('Error deleting profile picture:', problemDetails.detail)
-    throw new Error(`Delete failed: ${problemDetails.detail}`)
-  }
+  await fetchHandler(url, 'DELETE')
 }